feat: remove status and cron features from observatory
