--- src/observatory/src/constants.rs
@@ -1,2 +1,0 @@
-// 1 hour
-pub const CRON_INTERVAL_NS: u64 = 3_600_000_000_000;
