+++ src/observatory/src/types.rs
@@ -1,12 +1,6 @@
 pub mod state {
     use candid::{CandidType, Deserialize};
-    use junobuild_shared::types::cronjob::CronJobs;
-    use junobuild_shared::types::state::{
-        ArchiveTime, Controllers, MissionControlId, SegmentsStatuses, Timestamp, UserId, Version,
-    };
-    use std::collections::HashMap;
-
-    pub type CronTabs = HashMap<UserId, CronTab>;
+    use junobuild_shared::types::state::Controllers;
 
     #[derive(Default, Clone)]
     pub struct State {
@@ -16,28 +10,6 @@ pub mod state {
     #[derive(Default, CandidType, Deserialize, Clone)]
     pub struct StableState {
         pub controllers: Controllers,
-        pub cron_tabs: CronTabs,
-        pub archive: Archive,
-    }
-
-    #[derive(CandidType, Deserialize, Clone)]
-    pub struct CronTab {
-        pub mission_control_id: MissionControlId,
-        pub cron_jobs: CronJobs,
-        pub created_at: Timestamp,
-        pub updated_at: Timestamp,
-        pub version: Option<Version>,
-    }
-
-    #[derive(Default, CandidType, Deserialize, Clone)]
-    pub struct Archive {
-        pub statuses: HashMap<UserId, ArchiveStatuses>,
-    }
-
-    #[derive(CandidType, Deserialize, Clone)]
-    pub struct ArchiveStatuses {
-        pub timestamp: ArchiveTime,
-        pub statuses: Result<SegmentsStatuses, String>,
     }
 }
