+++ src/observatory/src/upgrade/impls.rs
@@ -1,0 +1,10 @@
+use crate::types::state::StableState;
+use crate::upgrade::types::upgrade::UpgradeStableState;
+
+impl From<&UpgradeStableState> for StableState {
+    fn from(state: &UpgradeStableState) -> Self {
+        StableState {
+            controllers: state.controllers.clone(),
+        }
+    }
+}
